Changes to Section 1.4
> p 7
>
> l 1.1: (C2) "linespacing" -> "line spacing".  Perhaps "linespacing" is
>     the correct way to write it in typographic texts, but it looks
>     wrong to me.

I personally do not like words like linespacing, pagenumbering, in manuals.
However, I also feel that it is unavoidable. I am not changing them here.


> l 3.1: (C4) "This is about 1.2x" -> "This is about 1.2 times"

Done

> l 3.2: (C5) "is default .72:.28" -> "is .72:.28 by default"

Done.

> paragraph 4: (?) I find this discussion of paragraphs a little
>     confusing (although I have some understanding of the fact that TeX
>     looks at line-related settings when it sees the \par at the end of
>     a paragraph).  This discussion may be too sophisticated for this
>     document.  

*** I think that this paragraph can be omiited here. I am enclosing it in
\start-stop-hiding.

>     Similarly, the final paragraph on the page assumes the
>     reader understands glue stretchability when it mentions that
>     \topskip is 11pt plus 55pt.

There is not much that we can do here, since the dimension is generated using
\the\topskip.

l 4.1: "Sometimes a lone does not have the maximum height or depth". I do not
        understand what this means. 

l 7-8: From the description, the difference between \baselineskip 
       and \lineskip is not clear. Why do we need \lineskip when we 
       already have \baselineskip (I need to reread the TeXbook).



> p 8
>
> (?) One of the examples uses the \start command without explanation.
>     I don't know what this does (but perhaps it is explained in an
>     earlier chapter of the overall manual).


I believe that it will be. 


> p 9
>
> l 3.3: (?) The text refers to "the last two quantities", which seems
>     to refer to bottom and line.  From context I think the text means
>     top and bottom

Changed.

l 3.1: "The height and depth determine the ration between the ratio between the
height and depth of a line". The beginning of the section suggests that the
height and depth depend on the characters in the line. This section does not
explain how to set \baselineskip and \lineskip.  

> l 5.5: (C5) "macropackages" -> "macro packages".

Done



git-svn-id: https://foundry.supelec.fr/svn/contextman@242 08d8c78c-c36f-4a2b-834e-415383aec5ca